Update. Переработка специфичных расширений
