+++ cs/Markdown/Readers/MarkdownToHtmlConverter.cs
@@ -1,7 +1,9 @@
-﻿using System.Text;
+﻿using System.Collections.Frozen;
+using System.Text;
 using Markdown.Converters;
 using Markdown.Converters.Base;
 using Markdown.Extensions;
+using Markdown.Helpers;
 using Markdown.Models;
 
 namespace Markdown.Readers;
@@ -17,21 +19,25 @@ public class MarkdownToHtmlConverter
 
     public string ConvertMarkdownToHtml(string text)
     {
+        ArgumentNullException.ThrowIfNull(text);
+
         var tokens = new List<TagPair>();
         foreach (var converter in converters)
         {
             tokens.AddRange(converter.Convert(text).ToList());
         }
 
-        var converterTokenPairs = tokens.FilterNonIntersectingTagPairs();
-        var replacementTokens = FilterBoldTokens(converterTokenPairs)
-            .FilterShielding(text);
+        var converterTokenPairs = TagPairHelpers.FilterNonIntersectingTagPairs(tokens);
+        var replacementTagPairs = FilterBoldTokens(converterTokenPairs);
+        var replacementTokens = TagPairHelpers.FilterShielding(replacementTagPairs, text).ToList();
         return RestoreToHtmlText(replacementTokens, text);
     }
 
-    private IEnumerable<TagPair> FilterBoldTokens(ICollection<TagPair> tokenPairs)
+    private IReadOnlyCollection<TagPair> FilterBoldTokens(IReadOnlyCollection<TagPair> tokenPairs)
     {
-        var convertTokens = tokenPairs.ToDictionaryByType();
+        var convertTokens = tokenPairs
+            .GroupBy(x => x.ConverterType)
+            .ToDictionary(x => x.Key, IReadOnlyCollection<TagPair> (x) => x.ToList());
         if (!convertTokens.ContainsKey(typeof(BoldConverter))
             || !convertTokens.ContainsKey(typeof(ItalicConverter)))
         {
@@ -41,11 +47,11 @@ public class MarkdownToHtmlConverter
         var boldTokens = convertTokens[typeof(BoldConverter)];
         var italicTokens = convertTokens[typeof(ItalicConverter)];
         convertTokens[typeof(BoldConverter)] = boldTokens.FilterOverlapping(italicTokens);
-        var boldFilterTokens = convertTokens.SelectMany(x => x.Value);
+        var boldFilterTokens = convertTokens.SelectMany(x => x.Value).ToList();
         return boldFilterTokens;
     }
 
-    private string RestoreToHtmlText(IEnumerable<TagToken> converterTokens, string originalText)
+    private string RestoreToHtmlText(IReadOnlyCollection<TagToken> converterTokens, string originalText)
     {
         var replacements = converterTokens.OrderBy(r => r.TagIndex);
         var resultBuilder = new StringBuilder();
@@ -57,9 +63,11 @@ public class MarkdownToHtmlConverter
             {
                 resultBuilder.Append(originalText.AsSpan(lastIndex, token.TagIndex - lastIndex));
             }
-
+        
             resultBuilder.Append(token.ReplaceTag);
-            lastIndex = token.TagIndex + token.SearchTag.Length;
+            lastIndex = token.TagIndex + token.SearchTag.Length >= originalText.Length 
+                ? originalText.Length
+                : token.TagIndex + token.SearchTag.Length;
         }
 
         if (lastIndex < originalText.Length)
